Change some class names to avoid clashes with proton4j names
changes the names of the classes to ClientX for now until a project name
is established and then they could be changed to ProjectX style for easy
identification in logs.  Adds some exception types that could be used
later to handle AMQP error conditions etc.  Adds some utility classes for
later use.