Refactor proton link base and associated code to handle state changes
Sender and Receiver need to refuse to perform various operations based
on the link state as well as that of the parent session and connection.
Refactor the link processing to create a quick mechanism to check this
state and reduce work on hot path methods like send and disposition
handling etc.