Refactor SequenceNumberMap and use it in unsettled delivery tracking
Use a fixed array with linked nodes to normalize access times to array
access values in cases where map load is small.  Use this map in the
unsettled delivery tacking to avoid the costs of the splay tree when
access tends to be forwards and back for delivery send and then
settlement received etc.