Adds some experimental Message types for Streamed message types
Some possible ways of allowing users to send partial messages or stream
data sections as data is available.  Also adds some support for custom
message formats beyond the default type defined in the AMQP
specification by allowing control of the contents that are encoded and
the message format used to populate the outbound delivery.