Fix several IO layer issues and refactor API and error usability
The IO layer needs to be singularly bound to the connection and not the
transport and the exception hierarchy should be easier to use and
determine which side of the connection local or remote was responsible
for the error.  Add many new tests for handling of errors on disconnect
of the connection or remote close of connection resources.