Add an exception type to indicate that the source is at EOF
For stream based and the buffer based decoder we need to indicate that a
new type read failed because the stream is at EOF and not for some other
malformed encoding type error.  More work should be done to ensure the
stream decoder throws when data that is expected to be in a stream is
not present.