Add tests and fixes for some stream receiver API uses
Ensure that credit is replenished and add improved handling of close of
body section input stream close when data read or not with checks for
Footer parse.  Tests that stream receiver can be used in a transaction
if desired.  Cleans up some APIs in the client top level package for
options etc.