PROTON-2524 Ensure that StreamReceiver refills the credit window
Stream receiver need to refill the credit window as deliveries are
returned from receive calls when they are already completed and when a
streamed delivery receives the final transfer as well.  Also auto accept
should be actively accepting deliveries that have completed.  The credit
window replenishment for the standard receiver and the stream receiver
impl should be mostly the same.