PROTON-2575 Add sub map and navigable map APIs to splay map
Provide a means of operating over a fixed range in the tracking
map for cases where dispositions arrive with a range of first
to last delivery ids.  This allows a more efficient handling of
these ranged dispositions and produces intermediate objects.
Also fixes an issue where the tracking map could be corrupted if
a remove and update operation falls within a very specific range
of values in the map tree.