PROTON-2661 Refactoring buffers and transports for netty 5 support
Changes the ProtonBuffer API to allow for compatability with both netty4
and netty5 and adds a transport in the proton client that will work with
the netty5 snapshot currently available.  Other refactorings and updates
internally to support the netty5 ownershup model for buffers and the
various differences between the two.