PROTON-2698 Add some additional API for checking on single values
Allow a test to check that a single expected value exists in what could
be a larger array of values such as offered or desired capabilities and
in sasl mechanisms offered.  Also fix an issue where connected event on
the peer client implementation was not getting triggered correctly in
all cases.